smartjpg2pptx/requirements.txt
python-pptx
opencv-python-headless
numpy
torch
torchvision
openai
comtypes
pypdfium2
reportlab
PyPDF2
